Moved all stats modal
