+++ ghost/admin/app/components/stats/modal-stats-all.hbs
